backend/services/gradcam_service.py
import numpy as np
import cv2
import tensorflow as tf
from tensorflow import keras
import sys
import json
import os
from PIL import Image
import matplotlib.pyplot as plt
import matplotlib.cm as cm

class GradCAMService:
    def __init__(self, model_path=None):
        """Inițializează serviciul Grad-CAM"""
        self.model = None
        self.last_conv_layer_name = None
        
        if model_path and os.path.exists(model_path):
            self.load_model(model_path)
        else:
            # Folosim un model pre-antrenat pentru demonstrație
            self.model = self.create_simple_model()
    
    def create_simple_model(self):
        """Creează un model simplu pentru demonstrație"""
        # Pentru demonstrație, folosim MobileNetV2 pre-antrenat
        base_model = keras.applications.MobileNetV2(
            input_shape=(224, 224, 3),
            include_top=False,
            weights='imagenet'
        )
        
        # Adăugăm layers pentru detecția deepfake
        model = keras.Sequential([
            base_model,
            keras.layers.GlobalAveragePooling2D(),
            keras.layers.Dense(128, activation='relu'),
            keras.layers.Dropout(0.2),
            keras.layers.Dense(1, activation='sigmoid')
        ])
        
        self.last_conv_layer_name = base_model.layers[-1].name
        return model
    
    def load_model(self, model_path):
        """Încarcă un model antrenat"""
        try:
            self.model = keras.models.load_model(model_path)
            # Găsește ultimul layer convolutional
            for layer in reversed(self.model.layers):
                if len(layer.output_shape) == 4:  # Conv layer
                    self.last_conv_layer_name = layer.name
                    break
        except Exception as e:
            print(f"Eroare la încărcarea modelului: {e}")
            self.model = self.create_simple_model()
    
    def preprocess_image(self, image_path, target_size=(224, 224)):
        """Preprocesează imaginea pentru model"""
        try:
            # Încarcă imaginea
            img = keras.preprocessing.image.load_img(image_path, target_size=target_size)
            img_array = keras.preprocessing.image.img_to_array(img)
            img_array = np.expand_dims(img_array, axis=0)
            img_array = keras.applications.mobilenet_v2.preprocess_input(img_array)
            return img_array, img
        except Exception as e:
            print(f"Eroare la preprocesarea imaginii: {e}")
            return None, None
    
    def make_gradcam_heatmap(self, img_array, pred_index=None, eps=1e-8):
        """Generează heatmap-ul Grad-CAM"""
        if self.model is None or self.last_conv_layer_name is None:
            raise ValueError("Model sau layer conv nu sunt inițializate")
        
        try:
            # Creează un model care mapează input-ul la activările conv layer-ului și la predicții
            grad_model = keras.models.Model(
                inputs=[self.model.inputs],
                outputs=[self.model.get_layer(self.last_conv_layer_name).output, self.model.output]
            )
            
            # Calculează gradientele
            with tf.GradientTape() as tape:
                last_conv_layer_output, preds = grad_model(img_array)
                if pred_index is None:
                    pred_index = tf.argmax(preds[0])
                class_channel = preds[:, pred_index]
            
            # Gradientele feature map-ului ultimului conv layer cu privire la clasa predicată
            grads = tape.gradient(class_channel, last_conv_layer_output)
            
            # Vector în care fiecare intrare este importanța medie a gradientului pe un anumit feature map
            pooled_grads = tf.reduce_mean(grads, axis=(0, 1, 2))
            
            # Înmulțește fiecare canal din feature map cu importanța lui
            last_conv_layer_output = last_conv_layer_output[0]
            heatmap = last_conv_layer_output @ pooled_grads[..., tf.newaxis]
            heatmap = tf.squeeze(heatmap)
            
            # Normalizează heatmap-ul între 0 și 1
            heatmap = tf.maximum(heatmap, 0) / tf.math.reduce_max(heatmap)
            return heatmap.numpy()
            
        except Exception as e:
            print(f"Eroare la generarea Grad-CAM: {e}")
            # Returnează un heatmap sintetic pentru debugging
            return self.generate_synthetic_heatmap()
    
    def generate_synthetic_heatmap(self, size=(7, 7)):
        """Generează un heatmap sintetic pentru demonstrație"""
        # Creează zone de activare în diferite regiuni
        heatmap = np.zeros(size)
        
        # Zonă de activare în centru (față)
        center_x, center_y = size[0] // 2, size[1] // 2
        heatmap[center_x-1:center_x+2, center_y-1:center_y+2] = 0.8
        
        # Zone de activare în colțuri (ochi, gură)
        heatmap[1, 1] = 0.6  # ochi stâng
        heatmap[1, -2] = 0.6  # ochi drept
        heatmap[-2, center_y] = 0.5  # zona gurii
        
        return heatmap
    
    def save_gradcam_heatmap(self, img_path, heatmap, output_path, alpha=0.6):
        """Salvează heatmap-ul suprapus pe imaginea originală"""
        try:
            # Încarcă imaginea originală
            img = keras.preprocessing.image.load_img(img_path)
            img = keras.preprocessing.image.img_to_array(img)
            
            # Redimensionează heatmap-ul la dimensiunea imaginii
            heatmap_resized = cv2.resize(heatmap, (img.shape[1], img.shape[0]))
            heatmap_resized = np.uint8(255 * heatmap_resized)
            
            # Aplică colormap
            heatmap_colored = cm.jet(heatmap_resized)
            heatmap_colored = np.uint8(255 * heatmap_colored)
            
            # Convertește la RGB
            heatmap_rgb = cv2.cvtColor(heatmap_colored, cv2.COLOR_RGBA2RGB)
            
            # Suprapune heatmap-ul pe imaginea originală
            superimposed_img = heatmap_rgb * alpha + img * (1 - alpha)
            superimposed_img = np.uint8(superimposed_img)
            
            # Salvează rezultatul
            result_img = Image.fromarray(superimposed_img)
            result_img.save(output_path, 'JPEG', quality=95)
            
            return True
            
        except Exception as e:
            print(f"Eroare la salvarea heatmap-ului: {e}")
            return False
    
    def generate_heatmap(self, image_path, output_path, confidence_score=None):
        """Funcția principală pentru generarea heatmap-ului"""
        try:
            # Preprocesează imaginea
            img_array, original_img = self.preprocess_image(image_path)
            if img_array is None:
                return False
            
            # Generează predicția
            prediction = self.model.predict(img_array, verbose=0)
            predicted_class = int(prediction[0][0] > 0.5)
            confidence = float(prediction[0][0])
            
            # Generează heatmap-ul Grad-CAM
            heatmap = self.make_gradcam_heatmap(img_array, predicted_class)
            
            # Salvează heatmap-ul suprapus
            success = self.save_gradcam_heatmap(image_path, heatmap, output_path)
            
            return {
                'success': success,
                'confidence': confidence,
                'predicted_class': predicted_class,
                'heatmap_path': output_path if success else None
            }
            
        except Exception as e:
            print(f"Eroare la generarea heatmap-ului: {e}")
            return {'success': False, 'error': str(e)}

def main():
    """Funcția principală care poate fi apelată din Node.js"""
    if len(sys.argv) < 3:
        print(json.dumps({'error': 'Parametri insuficienți'}))
        return
    
    image_path = sys.argv[1]
    output_path = sys.argv[2]
    confidence_score = float(sys.argv[3]) if len(sys.argv) > 3 else None
    
    # Inițializează serviciul
    gradcam = GradCAMService()
    
    # Generează heatmap-ul
    result = gradcam.generate_heatmap(image_path, output_path, confidence_score)
    
    # Returnează rezultatul ca JSON
    print(json.dumps(result))

if __name__ == "__main__":
    main()
